+++ include/mysql/service_thd_alloc.h
@@ -38,7 +38,7 @@ extern "C" {
 struct st_mysql_lex_string
 {
   char *str;
-  size_t length;
+  unsigned int length;
 };
 typedef struct st_mysql_lex_string MYSQL_LEX_STRING;
